+++ site/owid.scss
@@ -105,7 +105,7 @@
 @import "./DataInsightsNewsletterBanner.scss";
 @import "./DataPage.scss";
 @import "./DataPageContent.scss";
-@import "./multi-dim/MultiDimDataPageContent.scss";
+@import "./multiDim/MultiDimDataPageContent.scss";
 @import "./GrapherImage.scss";
 @import "./KeyDataTable.scss";
 @import "./MetadataSection.scss";
